Move SaiInterface to meta directory
