+++ lib/inc/ClientServerSai.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "SaiInterface.h"
+#include "meta/SaiInterface.h"
 
 #include <memory>
 #include <mutex>
