+++ lib/inc/RemoteSaiInterface.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "SaiInterface.h"
+#include "meta/SaiInterface.h"
 
 namespace sairedis
 {
