+++ lib/inc/ServerSai.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "SaiInterface.h"
+#include "meta/SaiInterface.h"
 #include "meta/SaiAttributeList.h"
 #include "syncd/SelectableChannel.h"
