+++ meta/DummySaiInterface.h
@@ -1,31 +1,31 @@
 #pragma once
 
-#include "lib/inc/SaiInterface.h"
+#include "SaiInterface.h"
 
 #include <memory>
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_REMOVE_ENTRY(ot)         \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_REMOVE_ENTRY(ot)          \
     virtual sai_status_t remove(                                    \
             _In_ const sai_ ## ot ## _t* ot) override;
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_CREATE_ENTRY(ot)         \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_CREATE_ENTRY(ot)          \
     virtual sai_status_t create(                                    \
             _In_ const sai_ ## ot ## _t* ot,                        \
             _In_ uint32_t attr_count,                               \
             _In_ const sai_attribute_t *attr_list) override;
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_SET_ENTRY(ot)            \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_SET_ENTRY(ot)             \
     virtual sai_status_t set(                                       \
             _In_ const sai_ ## ot ## _t* ot,                        \
             _In_ const sai_attribute_t *attr) override;
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_GET_ENTRY(ot)            \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_GET_ENTRY(ot)             \
     virtual sai_status_t get(                                       \
             _In_ const sai_ ## ot ## _t* ot,                        \
             _In_ uint32_t attr_count,                               \
             _Out_ sai_attribute_t *attr_list) override;
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_BULK_CREATE_ENTRY(ot)    \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_BULK_CREATE_ENTRY(ot)     \
     virtual sai_status_t bulkCreate(                                \
             _In_ uint32_t object_count,                             \
             _In_ const sai_ ## ot ## _t *ot,                        \
@@ -34,14 +34,14 @@
             _In_ sai_bulk_op_error_mode_t mode,                     \
             _Out_ sai_status_t *object_statuses) override;
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_BULK_REMOVE_ENTRY(ot)    \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_BULK_REMOVE_ENTRY(ot)     \
     virtual sai_status_t bulkRemove(                                \
             _In_ uint32_t object_count,                             \
             _In_ const sai_ ## ot ## _t *ot,                        \
             _In_ sai_bulk_op_error_mode_t mode,                     \
             _Out_ sai_status_t *object_statuses) override;
 
-#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_BULK_SET_ENTRY(ot)       \
+#define SAIMETA_DUMMYSAIINTERFACE_DECLARE_BULK_SET_ENTRY(ot)        \
     virtual sai_status_t bulkSet(                                   \
             _In_ uint32_t object_count,                             \
             _In_ const sai_ ## ot ## _t *ot,                        \
