+++ meta/SaiInterface.h
@@ -266,6 +266,5 @@ namespace sairedis
             virtual sai_status_t logSet(
                     _In_ sai_api_t api,
                     _In_ sai_log_level_t log_level) = 0;
-
     };
 }
