+++ syncd/HardReiniter.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "SaiInterface.h"
+#include "meta/SaiInterface.h"
 #include "SaiSwitch.h"
 #include "VirtualOidTranslator.h"
 #include "RedisClient.h"
